Moved algorithm decorators to individual files
