+++ vantage6-algorithm-tools/vantage6/algorithm/decorator/__init__.py
@@ -1,0 +1,10 @@
+from vantage6.algorithm.decorator.data import data
+from vantage6.algorithm.decorator.action import (
+    data_extraction,
+    pre_processing,
+    federated,
+    central,
+)
+from vantage6.algorithm.decorator.source_database import source_database
+from vantage6.algorithm.decorator.algorithm_client import algorithm_client
+from vantage6.algorithm.decorator.metadata import metadata
